refactor: relocate PresetSource and UpdateCompanyRequest to domain layer
Wave 3B: cleared 2 of 8 layer violations from audit-144 by moving
domain enums and input models out of synthorg.api.dto*.

PresetSource enum moved to synthorg/templates/preset_models.py.
synthorg/api/dto_personalities.py and templates/preset_service.py
import the enum from the templates layer.

UpdateCompanyRequest moved to synthorg/organization/models.py. The
API DTO module imports it back so HTTP-facing controller signatures
read unchanged. organization/services imports the domain model
directly with no api.dto_org dependency.

Remaining 6 violations (5 in providers/management, 1 in
settings/subscribers) are a larger refactor requiring coordination
with Litestar OpenAPI generation and dashboard codegen; tracked
separately. The patterns established here are the templates for
those follow-ups.
